+++ projects/iad-interface-admin/form/src/lib/dynamic-form/components/iad-form-input.component.ts
@@ -30,23 +30,16 @@ col-form-label">{{config.translate ? (config.label | translate) : config.label}}
                 (onBlur)="onBlur()">
               </p-inputMask>
             </ng-template>
-            <iad-tooltip-notifier *ngIf="!(config.readonly || config.disabled) && isInvalid" caption="Ошибка!" [text]="error">
-              <ng-template let-content="content" let-position="position"
-                           let-tooltipStyleClass="tooltipStyleClass" let-escape="escape">
-                <fa-icon
-                  [icon]="'exclamation-circle'" [size]="'2x'" [ngStyle]="{color: 'red'}"
-                  [iadPTooltip]="content" [tooltipPosition]="position" [tooltipStyleClass]="tooltipStyleClass" [escape]="false"
-                ></fa-icon>
+            <iad-tooltip-notifier *ngIf="!(config.readonly || config.disabled) && isInvalid"
+                                  caption="Ошибка!" [text]="error" [activated]="true">
+              <ng-template>
+                <fa-icon [icon]="'exclamation-circle'" [size]="'2x'" [ngStyle]="{color: 'red'}"></fa-icon>
               </ng-template>
             </iad-tooltip-notifier>
-            <iad-tooltip-notifier *ngIf="!(config.readonly || config.disabled) && !isInvalid" caption="Поле валидно!"
-                                  text="Валидация прошла успешно">
-              <ng-template let-content="content" let-position="position"
-                           let-tooltipStyleClass="tooltipStyleClass" let-escape="escape">
-                <fa-icon
-                  [icon]="'check-circle'" [size]="'2x'" [ngStyle]="{color: 'green'}"
-                  [iadPTooltip]="content" [tooltipPosition]="position" [tooltipStyleClass]="tooltipStyleClass" [escape]="false"
-                ></fa-icon>
+            <iad-tooltip-notifier *ngIf="!(config.readonly || config.disabled) && !isInvalid"
+                                  caption="Поле валидно!" text="Валидация прошла успешно" [activated]="true">
+              <ng-template>
+                <fa-icon [icon]="'check-circle'" [size]="'2x'" [ngStyle]="{color: 'green'}" ></fa-icon>
               </ng-template>
             </iad-tooltip-notifier>
           </div>